requirements.txt
sanic>=23.6.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
bcrypt>=4.1.0
PyJWT>=2.8.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
